requirements.txt
python-dotenv
psycopg2-binary
PyMuPDF
Pillow
pytesseract
